fix(app): fix magnetic module fixture conflicts
fixes the logic to surface a magnetic module fixture conflict if one exists and suppress conflicts
with staging area fixtures. changes the background color of the magnetic module item. splits out a
ModuleTable component for readability.

closes RAUT-913
